Island: Rename cc/resources/events.py -> cc/resources/agent_events.py
