+++ monkey/monkey_island/cc/resources/__init__.py
@@ -8,5 +8,5 @@ from .ip_addresses import IPAddresses
 from .agent_configuration import AgentConfiguration
 from .pba_file_upload import PBAFileUpload, LINUX_PBA_TYPE, WINDOWS_PBA_TYPE
 from .pba_file_download import PBAFileDownload
-from .events import AgentEvents
+from .agent_events import AgentEvents
 from .agents import Agents
